requirements.txt

PySide6>=6.5.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
tqdm>=4.66.0
pillow>=10.0.0
openai>=1.0.0
anthropic>=0.30.0
google-generativeai>=0.5.0
sherpa-onnx>=1.10.0
pyaudio>=0.2.14
keyboard>=0.13.5
numpy>=1.26.0,<2.0.0
pywin32>=306
pynvml>=11.5.0
GPUtil>=1.4.0
uiautomation>=2.0.0
pyperclip>=1.8.2
python-multipart>=0.0.9
qasync>=0.27.0
soundfile>=0.12.1
PyYAML>=6.0.1